pyproject.toml
[tool.poetry]
name = "iploc"
version = "0.1.0"
description = "Iploc is a CLI tool for finding the Geo Coordinates accosciated with an IP Address"
authors = ["Ryan Farrell <ryfarrell@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.4.2"
click = "^8.1.7"
pytest = "^7.4.2"
requests = "^2.31.0"
pytest-mock = "^3.11.1"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
